Android version

+++ app/src/main/assets/shaders/visualize.glsl
@@ -1,11 +1,13 @@
-precision highp float;
+precision mediump float;
 
 uniform sampler2D sampler;
 uniform vec3 bg_color;
 uniform vec3 fg_color;
 varying vec2 uv;
 
+
 #include "converse.glsl"
+
 
 void main() {
     vec4 l_c = texture2D(sampler, uv);
@@ -14,4 +16,5 @@ void main() {
     vec3 fill_color = mix(bg_color, fg_color, l);
     gl_FragColor = vec4(fill_color, 1.0);
 }
+
